move pkg/secrets to operator/pkg/secrets
(cherry picked from commit 187962a41588d883343c7a5c1b30e91fd58e476f)

# Conflicts:
#	operator/cmd/run/run.go
#	pkg/go.mod
#	pkg/go.sum
