+++ operator/pkg/secrets/secrets_test.go
